backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
psycopg2-binary==2.9.9
pgvector==0.2.4
pydantic-settings==2.1.0
pillow-heif==1.1.1
pillow-avif-plugin==1.4.3
sentence-transformers==5.2.0
geopy==2.4.1
opencv-python==4.13.0.90

# Background Tasks
celery==5.3.6
redis==5.0.1

# Testing dependencies
pytest==7.4.3
pytest-cov==4.1.0
httpx==0.26.0  # Required by TestClient

# Raw Image Processing
rawpy>=0.19.0
ExifRead>=3.0.0
PyExifTool==0.5.6

# Database dependencies
alembic==1.18.0